requirements.txt
adal==1.2.7
autopep8==2.0.1
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.12.0
azure-mgmt-core==1.3.2
azure-mgmt-monitor==5.0.1
azure-mgmt-resource==22.0.0
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
coloredlogs==15.0.1
cryptography==39.0.0
humanfriendly==10.0
idna==3.4
isodate==0.6.1
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
mypy-extensions==0.4.3
oauthlib==3.2.2
pathspec==0.10.3
platformdirs==2.6.2
portalocker==2.7.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
